requirements.txt
openai==1.51.2
pdfplumber==0.11.4
python-dotenv==1.0.1